#18 refact:actionを切り出し
